+++ godot/Model/NoteData.cs
@@ -12,14 +12,14 @@ public partial class NoteData : Resource
     public Array<Person> Tagged { get; set; }
 
     [Export]
-    public double UnixTime { get; set; }
+    public long UnixTime { get; set; }
 
     [Export]
     public bool Edited { get; set; }
 
     public string TimeString
     {
-        get => Time.GetTimeStringFromUnixTime((long)UnixTime);
+        get => Time.GetTimeStringFromUnixTime(UnixTime);
     }
     public string EditedString
     {
